% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{clipboard_context}
\alias{clipboard_context}
\alias{console_context}
\alias{custom_context}
\alias{markdown_context}
\alias{rstudio_context}
\title{clipboard_context}
\usage{
clipboard_context()

rstudio_context()

console_context()

markdown_context()

custom_context(output_mode = "console", nspc = 2, indent_context = 0,
  indent_head = TRUE)
}
\arguments{
\item{nspc}{The number of spaces for each indent level in the output context}

\item{indent_context}{The number of spaces applied initially to all lines in the output context}

\item{indent_head}{Logical. Apply the indent_context to the to the header row? Use FALSE if targetting cursor location.}
}
\value{
an output context. An input to _paste, _format, _contruct functions used to formatt whitespace.
}
\description{
clipboard_context

rstudio_context

console_context

markdown_context

custom_context
#' @param output_mode A named output mode, controls the target of the _paste functions options are "rstudioapi" or "console"
}

