% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_paste.R
\name{df_construct}
\alias{df_construct}
\title{df_construct}
\usage{
df_construct(input_table, oc = console_context())
}
\arguments{
\item{input_table}{an optional R object to parse instead of the clipboard.}

\item{oc}{an optional output context that defines the target and indentation.}
}
\value{
a character string containing the input formatted as a data.frame definition.
}
\description{
Parse the current clipboard as a table and return in data.frame format.
}
