% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_paste.R
\name{df_format}
\alias{df_format}
\title{df_format}
\usage{
df_format(input_table, output_context = clipboard_context())
}
\arguments{
\item{input_table}{an optional input tibble or data.frame to format.}

\item{output_context}{an optional output context that defines the target and indentation.}
}
\value{
nothing.
}
\description{
Parse the current clipboard as a table and paste to the clipboard in data.frame format.
}

