% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_paste.R
\name{df_paste}
\alias{df_paste}
\title{df_paste}
\usage{
df_paste()
}
\value{
the text pasted to the console. Useful for testing purposes.
}
\description{
Parse the current clipboard as a table and paste in at the cursor location in data.frame format.
}

