% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfdt_paste.R
\name{dfdt_construct}
\alias{dfdt_construct}
\title{dfdt_construct}
\usage{
dfdt_construct(input_table, oc = console_context(), class = NULL)
}
\arguments{
\item{input_table}{an optional R object to parse instead of the clipboard.}

\item{oc}{an optional output context that defines the target and indentation.}

\item{class}{either data.frame or data.table.}
}
\value{
a character string containing the input formatted as a data.frame definition.
}
\description{
Parse the current clipboard as a table and return in data.frame format.
}
