% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpasta.R
\name{dmdclip}
\alias{dmdclip}
\title{dmdclip}
\usage{
dmdclip(input)
}
\arguments{
\item{input}{a vector, data.frame, or tibble}
}
\value{
nothing
}
\description{
Formats input for presentation in mardkown as a preformatted chunk and inserts it onto the clipboard. Ready for pasting to Stack Overflow or Github.
}

