% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{dp_set_decimal_mark}
\alias{dp_set_decimal_mark}
\title{dp_set_decimal_mark}
\usage{
dp_set_decimal_mark(mark)
}
\arguments{
\item{mark}{The decimal mark to use when parsing "number" type data, as guessed by readr::guess_parser.}
}
\value{
NULL.
}
\description{
A function to optionally set the decimal mark if in a locale where it is not `.`. Will allow "3,14" to be parsed as 3.14, normally would be parsed as 314.
Will also handle spaces in numbers.
}

