% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{dp_set_max_rows}
\alias{dp_set_max_rows}
\title{dp_set_max_rows}
\usage{
dp_set_max_rows(num_rows)
}
\arguments{
\item{num_rows}{The number of rows of an input at which any of tribble_construct() or df_contruct() will abort parsing. Datapasta is untested on large tables. Use at own risk.}
}
\description{
dp_set_max_rows
}

