% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpasta.R
\name{dpasta}
\alias{dpasta}
\title{dpasta}
\usage{
dpasta(input)
}
\arguments{
\item{input}{a vector, data.frame, or tibble}
}
\value{
nothing
}
\description{
Formats input and inserts at either the current cursor or the console.
}
