% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfdt_paste.R
\name{dt_format}
\alias{dt_format}
\title{dt_format}
\usage{
dt_format(input_table, output_context = clipboard_context())
}
\arguments{
\item{input_table}{an optional input tibble or data.frame to format.}

\item{output_context}{an optional output context that defines the target and indentation.}
}
\value{
nothing.
}
\description{
Parse the current clipboard as a table and paste to the clipboard in data.table format.
}
