% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{escape_chars}
\alias{escape_chars}
\title{Title}
\usage{
escape_chars(char_vec)
}
\arguments{
\item{char_vec}{a character string with characters that may need escaping (eg. `"`)}
}
\value{
the original char_vec with an extra "\" inserted before each character that needs escaping.
}
\description{
Title
}

