% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_clip_paste.R
\name{eval_clip_dpasta}
\alias{eval_clip_dpasta}
\title{Evaluate the contents of the clipboard and paste the result}
\usage{
eval_clip_dpasta()
}
\value{
nothing.
}
\description{
Output format is determined by the class of the result.
}
