% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_clip_paste.R
\name{eval_clip_tribble}
\alias{eval_clip_tribble}
\title{Evaluate the contents of the clipboard and paste the result as a Tribble}
\usage{
eval_clip_tribble()
}
\value{
nothing.
}
\description{
Evaluate the contents of the clipboard and paste the result as a Tribble
}
