% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{guess_output_context}
\alias{guess_output_context}
\title{guess_output_context}
\usage{
guess_output_context()
}
\value{
a list containint the output target, space size of indent, and number of indents at context.
}
\description{
Return the a list containing the guessed output target context, either rstudio or the console.
}

