% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{guess_sep}
\alias{guess_sep}
\title{guess_sep}
\usage{
guess_sep(char_vec)
}
\arguments{
\item{char_vec}{a table from the clipboard in character vector form.}
}
\value{
the separator selected to parse char_vec as a table
}
\description{
Guesses the seprator based on a simple heuristic over the first 10 or less rows:
The separator chosen is the one that leads to the most columns, whilst parsing the same number of columns for each line (var=0).
The guessing algorithm ignores blank lines - which are lines that contain only the separator.
Options are in c(",","\\t","\\|,;")
}

