% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{nchar_type}
\alias{nchar_type}
\title{nchar_type}
\usage{
nchar_type(df_col_row, df_col_type)
}
\arguments{
\item{df_col_row}{a character string}

\item{df_col_type}{the type the string will be converted to.}
}
\value{
The number of characters wide this data would be in when rendered in text
}
\description{
nchar_type
}
