% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{nquote_str}
\alias{nquote_str}
\title{Count the number of quotes in a string}
\usage{
nquote_str(char_vec)
}
\arguments{
\item{char_vec}{the string to count quotes in}
}
\value{
a number, possibly 0.
}
\description{
Count the number of quotes in a string
}
