% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{pad_to}
\alias{pad_to}
\title{pad_to}
\usage{
pad_to(char_vec, char_length)
}
\arguments{
\item{char_vec}{character vector.}

\item{char_length}{length to pad to.}
}
\value{
char_vec left-padded with spaces to char_length.
}
\description{
Left pad string to a certain size. A helper function for getting spacing in table correct.
}
