% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{read_clip_tbl_guess}
\alias{read_clip_tbl_guess}
\title{read_clip_table_guess}
\usage{
read_clip_tbl_guess(x = NULL, ...)
}
\arguments{
\item{x}{contents of the clipboard}

\item{...}{arguments passed to read.table}
}
\value{
a parsed table from the clipboard. Separator is guessed.
}
\description{
Similar to read_clip_tbl from clipr,
however it will error if there are less than 2 rows
and it tries to guess the separator.
}
