% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_paste.R
\name{render_type}
\alias{render_type}
\title{render_type}
\usage{
render_type(char_vec, type_str)
}
\value{
A vector parsed from the clipboard as ether a character string or a
character vector. The type attribute contains the type guessed by `readr`.
}
\description{
Renders a character vector as R types for pasting into Rstudio.
Strings are quoted. Numbers, NaN, NA, logicals etc are not.
}

