% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{render_type_pad_to}
\alias{render_type_pad_to}
\title{render_type_pad_to}
\usage{
render_type_pad_to(char_vec, char_type, char_length)
}
\arguments{
\item{char_vec}{a character vector}

\item{char_type}{a string type from readr::guess_parser}

\item{char_length}{a string length to pad to.}
}
\value{
a string containing the representation of char_vec as char_type in the RStudio source editor,
left-padded with spaces to char_length.
}
\description{
Based on a type and length, render a character string as the type in text.
Pad to the desired length.
}

