% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfdt_paste.R
\name{tortellini}
\alias{tortellini}
\title{wrap the datapasta around itself}
\usage{
tortellini(col_struct, defn_width = 80, indent_context = 0, add_comma = TRUE)
}
\arguments{
\item{col_struct}{input structure - a split apart column definition}

\item{defn_width}{total number of characters in a line (includes column name and indent on line 1)}

\item{indent_context}{the level of indent in spaces in the current editor pane}

\item{add_comma}{add one final comma to the end of the wrapped column def? Useful when pasting together columns.}
}
\value{
w wrapped string
}
\description{
wrap the datapasta around itself
}
