% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{tribble_construct}
\alias{tribble_construct}
\title{tribble_construct}
\usage{
tribble_construct(input_table, oc = console_context())
}
\arguments{
\item{input_table}{an optional input `data.frame`. If `input_table` is supplied, then nothing is read from the clipboard.}

\item{oc}{an optional output context that defines the target and indentation. Default is console.
Table is output as `tribble()` call. Useful for creating reproducible examples.}
}
\value{
The parsed table text.
}
\description{
Parse the current clipboard as a table, or use the table argument supplied, and return as a character string.
}
