% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{tribble_format}
\alias{tribble_format}
\title{tribble_format}
\usage{
tribble_format(input_table, output_context = console_context())
}
\arguments{
\item{input_table}{an optional input `data.frame`. If `input_table` is supplied, then nothing is read from the clipboard.}

\item{output_context}{an optional output context that defines the target and indentation. Default is console.
Table is output as `tribble()` call. Useful for creating reproducible examples.}
}
\value{
Nothing.
}
\description{
Parse the current clipboard as a table, or use the table argument supplied, and paste to the clipboard in tribble format.
}
