% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_paste.R
\name{vector_construct}
\alias{vector_construct}
\title{vector_construct}
\usage{
vector_construct(input_vector, oc = console_context())
}
\arguments{
\item{input_vector}{An input vector to be formatted for output. If supplied, no data is read from the clipboard.}

\item{oc}{an optional output context that defines the output indentation.}
}
\value{
A string containing the input formatted as a vector definition.
}
\description{
Returns a formatted character string, either from clipboard or supplied argument, as a vector definition. Considers `,`, `tab`, `newline` as delimiters.
If a single character string is passed as an argument, it will be split to form a vector.
}
