% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_paste.R
\name{vector_format}
\alias{vector_format}
\title{vector_format}
\usage{
vector_format(input_vector, output_context = console_context())
}
\arguments{
\item{input_vector}{An input vector to be formatted for output. If supplied, no data is read from the clipboard.}

\item{output_context}{an optional output context that defines the output indentation.}
}
\value{
nothing.
}
\description{
Writes data to the clipboard, either from clipboard or supplied argument. Writes a horizontally formatted character vector on
a single line. Considers `,`, `tab`, `newline` as delimiters. If a single character string is passed as an argument, it will be split to form a vector.
}
