% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_paste.R
\name{vector_format_vertical}
\alias{vector_format_vertical}
\title{vector_format_vertical}
\usage{
vector_format_vertical(input_vector, output_context = clipboard_context())
}
\arguments{
\item{input_vector}{An input vector to be formatted for output. If supplied, no data is read from the clipboard.}

\item{output_context}{an optional output context that defines the output target and indentation.
The default behaviour is target the rstudioapi and fall back to console if it is not available.}
}
\value{
nothing.
}
\description{
Writes data to clipboard, either from clipboard or supplied argument, as a vertically formatted character vector over many lines.
Considers `,`, `tab`, `newline` as delimeters. If a single character string is passed as an argument, it will be split to form a vector.
}

