% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfiddle.R
\name{zzz_rs_dfiddle}
\alias{zzz_rs_dfiddle}
\title{dfiddle}
\usage{
zzz_rs_dfiddle()
}
\value{
a fiddled version of your selection (invisibly)
}
\description{
An addin to fiddle your RStudio selections to better things.
Make a selection in RStudio and dfiddle will update it inline. Good for:
Converting Text to vectors (`c()`), pivoting between horizontal and vertical vectors,
reflowing tribble() and data.frame() definitions to have nice indenting and padding.
}
