% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfiddle.R
\name{zzz_rs_toggle_quotes}
\alias{zzz_rs_toggle_quotes}
\title{Toggle Quotes}
\usage{
zzz_rs_toggle_quotes()
}
\value{
The toggled vector (invisibly).
}
\description{
An addin to toggle between quotes and bare vectors. Applies to a vector selected
in an RStudio source editor. Works with horizontal or vertical form.
}
