% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_paste.R
\name{df_paste}
\alias{df_paste}
\title{df_paste}
\usage{
df_paste(input_table, output_context = guess_output_context())
}
\arguments{
\item{input_table}{an optional input tibble or data.frame to format.}

\item{output_context}{an optional output context that defines the target and indentation.
The default behaviour is target the rstudioapi and fall back to console if it is not available.}
}
\value{
nothing.
}
\description{
Parse either: the current clipboard, or a supplied argument, as a table and paste in at the cursor location in data.frame format.
}

