% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{render_type}
\alias{render_type}
\title{render_type}
\usage{
render_type(char_vec, char_type)
}
\arguments{
\item{char_vec}{a character vector containing text to be rendered as the type indicated by type_str}

\item{char_type}{a string describing the type of char_vec}
}
\value{
A vector parsed from the clipboard as ether a character string or a
character vector. The type attribute contains the type guessed by `readr`.
}
\description{
Renders a character vector as R types for pasting into Rstudio.
Strings are quoted. Numbers, NA, logicals etc are not.
}
