% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_paste.R
\name{tortellini}
\alias{tortellini}
\title{wrap the datapasta around itself}
\usage{
tortellini(s, n = 80, indent_context = 0, add_comma = TRUE)
}
\arguments{
\item{s}{input string}

\item{n}{number of characters for text (includes column name on line 1)}

\item{indent_context}{the level of indent in spaces in the current editor pane}

\item{add_comma}{add one final comma to the end of the wrapped column def? Useful when pasting together columns.}
}
\value{
w wrapped string
}
\description{
wrap the datapasta around itself
}
