% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble_paste.R
\name{tribble_paste}
\alias{tribble_paste}
\title{tribble_paste}
\usage{
tribble_paste()
}
\value{
The parsed table text. Useful for testing.
}
\description{
Parse the current clipboard as a table and paste in at the cursor location in tribbble format.
}

