% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_paste.R
\name{vector_paste}
\alias{vector_paste}
\title{vector_paste}
\usage{
vector_paste()
}
\value{
nothing.
}
\description{
Pastes data from clipboard as a horizontally formatted character vector on
a single line. Considers , | tab newline as delimeters.
}

