% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_paste.R
\name{vector_paste_vertical}
\alias{vector_paste_vertical}
\title{vector_paste_vertical}
\usage{
vector_paste_vertical()
}
\value{
nothing.
}
\description{
Pastes data from clipboard as a vertically formatted character vector on
a multiple lines. One line is used per element. Considers , | tab newline as delimeters.
}

